Codechange: move main function(s) to separate files
