+++ src/os/macosx/CMakeLists.txt
@@ -4,6 +4,7 @@ add_files(
     font_osx.h
     macos.h
     macos.mm
+    osx_main.cpp
     osx_stdafx.h
     string_osx.cpp
     string_osx.h
