+++ src/os/os2/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_files(
     os2.cpp
+    os2_main.cpp
     CONDITION OPTION_OS2
 )
