+++ src/os/os2/os2.cpp
@@ -12,8 +12,6 @@
 #include "../../gui.h"
 #include "../../fileio_func.h"
 #include "../../fios.h"
-#include "../../openttd.h"
-#include "../../core/random_func.hpp"
 #include "../../string_func.h"
 #include "../../textbuf_gui.h"
 #include "../../thread.h"
@@ -167,16 +165,6 @@ void ShowOSErrorBox(const char *buf, bool system)
 	/* terminate PM env. */
 	WinDestroyMsgQueue(hmq);
 	WinTerminate(hab);
-}
-
-int CDECL main(int argc, char *argv[])
-{
-	SetRandomSeed(time(nullptr));
-
-	/* Make sure our arguments contain only valid UTF-8 characters. */
-	for (int i = 0; i < argc; i++) StrMakeValidInPlace(argv[i]);
-
-	return openttd_main(argc, argv);
 }
 
 bool GetClipboardContents(char *buffer, const char *last)
