+++ src/os/unix/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_files(
     crashlog_unix.cpp
+    unix_main.cpp
     CONDITION UNIX AND NOT APPLE AND NOT OPTION_OS2
 )
